+++ src/test/kotlin/com/zepben/evolve/cim/iec61970/base/core/IdentifiedObjectTest.kt
@@ -45,49 +45,133 @@ internal class IdentifiedObjectTest {
     }
 
     @Test
-    internal fun names() {
+    internal fun `User can add names to identified object`() {
+
         val identifiedObject = object : IdentifiedObject("id") {}
         val nameType = NameType("type")
-
         assertThat(identifiedObject.numNames(), equalTo(0))
 
         identifiedObject.addName(nameType, "1")
-        identifiedObject.addName(nameType, "2")
-        identifiedObject.addName(nameType, "3")
-        assertThat(identifiedObject.numNames(), equalTo(3))
+
+        assertThat(identifiedObject.numNames(), equalTo(1))
+    }
+
+    @Test
+    internal fun `Adding identical name to the same object doesn't change the list of names belonging to it`() {
+        val (identifiedObject: IdentifiedObject, nameType: NameType) = createMultipleBaseNames()
+        val originalNames = identifiedObject.names
+
         identifiedObject.addName(nameType, "1")
+        val newNames = identifiedObject.names
+        assertThat(originalNames, equalTo(newNames))
+    }
+
+    @Test
+    internal fun `getName obtain expected Name object`() {
+        val (identifiedObject: IdentifiedObject, nameType: NameType) = createMultipleBaseNames()
 
         val name1 = identifiedObject.getName("type", "1")!!
         val name2 = identifiedObject.getName("type", "2")!!
         val name3 = identifiedObject.getName("type", "3")!!
 
+        assertThat(name2.name, equalTo("2"))
+        assertThat(name2.type.name, equalTo("type"))
+
+        // Make sure item obtained are different
         assertThat(name1, not(equalTo(name2)))
         assertThat(name1, not(equalTo(name3)))
         assertThat(name2, not(equalTo(name3)))
+    }
 
-        assertThat(identifiedObject.numNames(), equalTo(3))
+    @Test
+    internal fun `getName(String, String) and getName(NameType, String) grabs the same object`() {
+        val (identifiedObject: IdentifiedObject, nameType: NameType) = createMultipleBaseNames()
+
+        val name2 = identifiedObject.getName("type", "2")!!
+        val dupeName2 = identifiedObject.getName(nameType, "2")
+        assertThat(name2, sameInstance(dupeName2))
+    }
+
+    @Test
+    internal fun `nameType contains the names added to the identified object`() {
+        val (identifiedObject: IdentifiedObject, nameType: NameType) = createMultipleBaseNames()
+
+        assertThat(nameType.hasName("1"), equalTo(true))
+        assertThat(nameType.hasName("2"), equalTo(true))
+        assertThat(nameType.hasName("3"), equalTo(true))
+    }
+
+    @Test
+    internal fun `removeName removes name from the identified object and the nameType`() {
+        val (identifiedObject: IdentifiedObject, nameType: NameType) = createMultipleBaseNames()
+        val name1 = identifiedObject.getName("type", "1")!!
+        val name2 = identifiedObject.getName("type", "2")!!
+        val name3 = identifiedObject.getName("type", "3")!!
 
         assertThat(identifiedObject.removeName(name1), equalTo(true))
         assertThat(identifiedObject.removeName(name1), equalTo(false))
         assertThat(identifiedObject.removeName(null), equalTo(false))
         assertThat(identifiedObject.numNames(), equalTo(2))
-
-        assertThat(identifiedObject.getName("type", name2.name), equalTo(name2))
+        assertThat(nameType.hasName("1"), equalTo(false))
 
         assertThat(identifiedObject.names, containsInAnyOrder(name2, name3))
+    }
+
+    @Test
+    internal fun `clearNames removes all names from the identified object and the nameType`() {
+        val (identifiedObject: IdentifiedObject, nameType: NameType) = createMultipleBaseNames()
+
+        assertThat(identifiedObject.numNames(), equalTo(3))
+        assertThat(nameType.hasName("1"), equalTo(true))
+        assertThat(nameType.hasName("2"), equalTo(true))
+        assertThat(nameType.hasName("3"), equalTo(true))
 
         identifiedObject.clearNames()
         assertThat(identifiedObject.numNames(), equalTo(0))
+        assertThat(nameType.hasName("1"), equalTo(false))
+        assertThat(nameType.hasName("2"), equalTo(false))
+        assertThat(nameType.hasName("3"), equalTo(false))
+    }
 
-        // Make sure you can add an item back after it has been removed
+    @Test
+    internal fun `user can add the same name back after it has been removed`(){
+        val identifiedObject = object : IdentifiedObject("id") {}
+        val nameType = NameType("type")
+
         identifiedObject.addName(nameType, "1")
+        val name1 = identifiedObject.getName("type", "1")
+        identifiedObject.clearNames()
+
+        identifiedObject.removeName(name1)
+        assertThat(identifiedObject.numNames(), equalTo(0))
+    }
+
+    @Test
+    internal fun `removing name from empty name list does not cause any issue`(){
+        val identifiedObject = object : IdentifiedObject("id") {}
+        val nameType = NameType("type")
+
+        identifiedObject.addName(nameType, "1")
+        val name1 = identifiedObject.getName("type", "1")
         assertThat(identifiedObject.numNames(), equalTo(1))
 
         identifiedObject.removeName(name1)
         assertThat(identifiedObject.numNames(), equalTo(0))
 
-        // Make sure you can call remove on an empty list.
-        identifiedObject.removeName(name2)
-        assertThat(identifiedObject.numNames(), equalTo(0))
+        identifiedObject.addName(nameType, "1")
+        val dupeName1 = identifiedObject.getName(nameType, "1")
+        assertThat(name1, not(sameInstance(dupeName1)))
+        assertThat(identifiedObject.numNames(), equalTo(1))
+    }
+
+    private fun createMultipleBaseNames(): Pair<IdentifiedObject, NameType> {
+        val identifiedObject = object : IdentifiedObject("id") {}
+        val nameType = NameType("type")
+
+        identifiedObject.addName(nameType, "1")
+        identifiedObject.addName(nameType, "2")
+        identifiedObject.addName(nameType, "3")
+
+        return identifiedObject to nameType
     }
 }